[refactor] main : detach parsing part
